+++ mobile/appcontainer/VersionBar.js
@@ -24,7 +24,7 @@ export class VersionBar extends Component {
     render() {
         const env = XH.getEnv('appEnvironment'),
             version = XH.getEnv('clientVersion'),
-            isVisible = (env !== 'Production' || XH.getPref('xhForceEnvironmentFooter'));
+            isVisible = (env !== 'Production' || XH.getUser().isHoistAdmin || XH.getPref('xhForceEnvironmentFooter'));
 
         if (!isVisible) return null;