+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Upcoming Release
 
+### 🎁 New Features
+* `Panel.tbar` and `Panel.bbar` now accept an array of Elements and will auto-generate a toolbar
+  to contain them. 
+
 ### 🐞 Bug Fixes
 * `Select` will now let the user edit existing text in conditions where it is expected to be 
   editable. #880