+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 # Changelog
 
-## v22.0.0-SNAPSHOT (under development)
+## Upcoming Release
+
+### 🎁 New Features
+* `Panel.tbar` and `Panel.bbar` now accept an array of Elements and will auto-generate a toolbar
+  to contain them. 
+* Hoist Admins now always see the VersionBar in the footer.
+
+### 🐞 Bug Fixes
+* `Select` will now let the user edit existing text in conditions where it is expected to be 
+  editable. #880
+* The Admin "Config Differ" tool has been updated to reflect changes to `Record` made in v22. It is
+  once again able to apply remote config values.
+
+## v22.0.0 - 2019-04-29
 
 ### 🎁 New Features